requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
transformers>=4.30.0
huggingface-hub>=0.15.0
pyyaml>=6.0
tqdm>=4.65.0
scipy>=1.10.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
opencv-python>=4.8.0
pillow>=10.0.0
einops>=0.6.1
timm>=0.9.0
